Fix use latest fog-core/fog-json with Gemfile-edge
Move fog-core and fog-json to be in the Gemfile to match the
Gemfile-edge and ensure that resulting combination of dependencies will
be resolved either with the latest from git or pinned and not a mixture
of the two.

Declaring the same dependencies in Gemfile and gemspec will not result
in one overriding the other, instead need to maintain the same
dependencies in equivalent files.