Fix use latest fog-core/fog-json with Gemfile-edge
Move the git based versions of fog-core and fog-json and pinned
releases to the same Gemfile with logic to determine which to use based
on an env variable so that only one Gemfile need be maintained.

Declaring the same dependencies in Gemfile and gemspec will not result
in one overriding the other, instead need to maintain the two sets of
dependencies in the same file.